requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
prisma==0.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.7.1
python-dotenv==1.0.1